reformat project, removing all package file from root
