+++ Makefile
@@ -27,17 +27,7 @@ nuke:
 
 ## Build the plugin binary
 build:
-	CGO_ENABLED=0 go build -ldflags $(LDFLAGS) cmd/coredns.go
-
-
-## Generate new helm package and update chart yaml file
-helm-update:
-	helm package charts/k8s-gateway -d charts.tmp/charts
-	helm repo index --url https://ori-edge.github.io/k8s_gateway/ --merge index.yaml charts.tmp/
-	mv charts.tmp/charts/* charts/
-	mv charts.tmp/index.yaml .
-	rmdir charts.tmp/charts
-	rmdir charts.tmp
+	CGO_ENABLED=0 go build -ldflags $(LDFLAGS) main.go
 
 .PHONY: test
 test:
